+++ src/components/CredentialsForm.jsx
@@ -1,8 +1,8 @@
-import { useState, useEffect } from "react";
+import { useEffect } from "react";
 import { useFormik } from "formik";
 import * as Yup from "yup";
 
-import { Button } from "@/components/ui/button";
+import { Button } from "@/components/ui/Button.jsx";
 import {
   Sheet,
   SheetClose,
@@ -11,7 +11,7 @@ import {
   SheetHeader,
   SheetTitle,
   SheetTrigger,
-} from "@/components/ui/sheet";
+} from "@/components/ui/Sheet.jsx";
 
 export function CredentialsForm({ onSubmit, initialValues }) {
   useEffect(() => {
