+++ src/components/DisplayResult.jsx
