+++ src/pages/index.jsx
@@ -1,4 +1,4 @@
-import { Button } from "@/components/ui/button";
+import { Button } from "@/components/ui/Button.jsx";
 import { Link } from "react-router-dom";
 
 export function HomePage() {
